+++ .gitignore
@@ -4,3 +4,4 @@
 3-main.py
 set_up.sql
 .env
+.dev
