+++ client/src/App.tsx
@@ -7,6 +7,7 @@ function App() {
       <div className="flex space-x-4">
         <Link className="border border-secondary-purple p-2 w-fit rounded-lg" to="/register">Register page</Link>
         <Link className="border border-secondary-purple p-2 w-fit rounded-lg" to="/login">Login page</Link>
+        <Link className="border border-secondary-purple p-2 w-fit rounded-lg" to="/conversations">Conversations page</Link>        
       </div>
 
     </>
