--- client/src/components/list/chatList/ChatList.jsx
@@ -1,70 +1,0 @@
-import React, { useState } from 'react'
-import "./chatlist.css"
-const Chatlist = () => {
-  const [addMode,setAddMode]= useState(false)
-  return (
-    <div className='chatList'>
-      <div className='header'>
-        <div className='logoSection'>
-            <img src='./talkify.png' alt='Talkify Logo' />
-            <div class="addSymbol">
-                <img src="./plus.png" alt="" />
-            </div>
-            </div>
-        <p className='chatTitle'>Chats</p>
-        </div>
-      <div className="search">
-        <div className="searchBar">
-          <img src="./searchblack.png" alt="" />
-          <input type="text" placeholder="Search" />
-        </div>
-        </div>
-        <div className="item">
-          <img src="./user.png" alt=""/>
-          <div className="texts">
-            <span>Julie Li</span>
-            <p>Hey there</p>
-      </div>
-      </div>
-      <div className="item">
-          <img src="./user.png" alt=""/>
-          <div className="texts">
-            <span>Julie Li</span>
-            <p>Hey there</p>
-      </div>
-      </div>
-
-      <div className="item">
-          <img src="./user.png" alt=""/>
-          <div className="texts">
-            <span>Julie Li</span>
-            <p>Hey there</p>
-      </div>
-      </div>
-
-      <div className="item">
-          <img src="./user.png" alt=""/>
-          <div className="texts">
-            <span>Julie Li</span>
-            <p>Hey there</p>
-      </div>
-      </div>
-      <div className="item">
-          <img src="./user.png" alt=""/>
-          <div className="texts">
-            <span>Julie Li</span>
-            <p>Hey there</p>
-      </div>
-      </div>
-      <div className="item">
-          <img src="./user.png" alt=""/>
-          <div className="texts">
-            <span>Julie Li</span>
-            <p>Hey there</p>
-      </div>
-      </div>
-    </div>
-  )
-}
-
-export default Chatlist
