--- client/src/routes/conversations/[id].tsx
@@ -1,8 +1,0 @@
-import { useParams } from "react-router-dom";
-
-const Conversation = () => {
-  const { id } = useParams<{ id: string }>();
-  return <div>Conversation with id {id}</div>;
-};
-
-export default Conversation;
