+++ client/src/routes/conversations/components/Conversation.tsx
@@ -1,0 +1,8 @@
+import Chat from '../../../components/chat/Chat'
+const Conversation = () => {
+  return (
+    <Chat />
+  )
+};
+
+export default Conversation;
